This is commit number 12; brings me up to play game again on schedule
Unfortunately, I did not deploy as often as I would have liked in the
last couple of days because I have not been looking at the suggested
schedule and instead, I've been running through what my brain suggests
should be the functionality of this app. Please forgive me, future me.

I have now managed to check for winner, declare the winner, and display
who's turn it is for the user.
This was all done in games/events.js intially.
Then I made a PATCH request to the API to update the game after a win or
loss and add messaging for when the game is over. And the user can no
longer play on the board. This is done in the events.js, the api.js, the
ui.js.

I also changed the background color in the index.scss file and
and I've addec my GET request curl script in the index.sh